Move add_round route to new folder /routes/add_round.
