+++ tests/test_add_round.py
@@ -3,11 +3,11 @@ from globals.globals import app
 from unittest import mock
 from unittest.mock import MagicMock
 import datetime
-from routes.rounds.add_round import (validate_date_interval,
-                                     validate_dates,
-                                     validate_start_before_end,
-                                     validate_name_length,
-                                     validate_unique_name)
+from routes.add_round.add_round import (validate_date_interval,
+                                        validate_dates,
+                                        validate_start_before_end,
+                                        validate_name_length,
+                                        validate_unique_name)
 from dateutil import parser
 
 
@@ -48,13 +48,13 @@ class TestAddRound(unittest.TestCase):
         self.assertFalse(validate_unique_name("unqiuename"))
 
     @mock.patch('flask_login.utils._get_user')
-    @mock.patch('routes.rounds.add_round.db.session.add', MagicMock(return_value=None))
-    @mock.patch('routes.rounds.add_round.db.session.commit', MagicMock(return_value=None))
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.db.session.add', MagicMock(return_value=None))
+    @mock.patch('routes.add_round.add_round.db.session.commit', MagicMock(return_value=None))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=True))
     def test_add_round_valid_round_logged_in(self, current_user):
         current_user.is_authenticated = True
         with self.client as client:
@@ -62,13 +62,13 @@ class TestAddRound(unittest.TestCase):
                 '/rounds/add', data=dict(name="test", start_date="01/01/2001 12:00 AM", end_date="02/02/2002 12:00 PM"), follow_redirects=True)
             self.assertIn(b'All Rounds', response.data)
 
-    @mock.patch('routes.rounds.add_round.db.session.add', MagicMock(return_value=None))
-    @mock.patch('routes.rounds.add_round.db.session.commit', MagicMock(return_value=None))
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.db.session.add', MagicMock(return_value=None))
+    @mock.patch('routes.add_round.add_round.db.session.commit', MagicMock(return_value=None))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=True))
     def test_add_round_valid_round_not_logged_in(self):
         with self.client as client:
             response = client.post(
@@ -76,11 +76,11 @@ class TestAddRound(unittest.TestCase):
             self.assertIn(b'Sign In', response.data)
 
     @mock.patch('flask_login.utils._get_user')
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=False))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=False))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=True))
     def test_add_round_date_interval_validation_invalid_date_interval_flash(self, current_user):
         current_user.is_authenticated = True
         with self.client as client:
@@ -90,11 +90,11 @@ class TestAddRound(unittest.TestCase):
                 b'Failed to create new round. A round already exists between the given dates, please enter valid dates.', response.data)
 
     @mock.patch('flask_login.utils._get_user')
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=False))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=False))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=True))
     def test_add_round_date_in_past_validation_dates_in_past_flash(self, current_user):
         current_user.is_authenticated = True
         with self.client as client:
@@ -104,11 +104,11 @@ class TestAddRound(unittest.TestCase):
                 b'Failed to create new round. One or more of the supplied dates are in the past, please enter valid dates.', response.data)
 
     @mock.patch('flask_login.utils._get_user')
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=False))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=False))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=True))
     def test_add_round_start_date_after_end_date_validation_flash(self, current_user):
         current_user.is_authenticated = True
         with self.client as client:
@@ -118,11 +118,11 @@ class TestAddRound(unittest.TestCase):
                 b'Failed to create new round. A round cannot be created with an end date before the start date, please enter valid dates.', response.data)
 
     @mock.patch('flask_login.utils._get_user')
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=False))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=False))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=True))
     def test_add_round_name_length_greater_than_max_validation_flash(self, current_user):
         current_user.is_authenticated = True
         with self.client as client:
@@ -132,11 +132,11 @@ class TestAddRound(unittest.TestCase):
                 b'Failed to create new round. The maximum name length is 12 characters.', response.data)
 
     @mock.patch('flask_login.utils._get_user')
-    @mock.patch('routes.rounds.add_round.validate_date_interval', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_dates', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_start_before_end', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_name_length', MagicMock(return_value=True))
-    @mock.patch('routes.rounds.add_round.validate_unique_name', MagicMock(return_value=False))
+    @mock.patch('routes.add_round.add_round.validate_date_interval', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_dates', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_start_before_end', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_name_length', MagicMock(return_value=True))
+    @mock.patch('routes.add_round.add_round.validate_unique_name', MagicMock(return_value=False))
     def test_add_round_unique_name_validation_flash(self, current_user):
         current_user.is_authenticated = True
         with self.client as client:
